Add to readme an indication that the app is known as "explorook"
This may seem obvious to some, but nowhere on the page is the term used. One would have to notice it in the project's name. But since we may be brought here from within the tool or site docs, it just seems reasonable to at least make mention of the tool's codename/project name. 

If you disagree, I'll understand. That's why I've broken this out as its own commit.