add info on installing the desktop app
The readme currently has no discussion at all of installation. While it's readily offered from within the Rookout UI, if you follow the right workflow, not everyone seeing this page would realize that, so it seemed worth mentioning.

Also, I put it in a new section that I put down near the bottom. If you think it more appropriate to move higher, feel free of course. I just didn't want to presume you'd care for that, if you care for this change at all. :-)

Finally, note that I'd found the video I offer on [a page in the RookOut docs](https://docs.rookout.com/docs/source-repos/#local-filesystem-rookout-desktop-app). And you may notice I've offered the URL "bare" (not as a hyperlink, even), just dropping in the video url onto the page. That's not a mistake: various resources indicate that github now will automatically render the HTML needed for showing that video automatically for us. I did NOT find that to be the case when I used the "preview" feature here in the github UI, but it may be that the conversion of the link to such HTML happens in some processing that github does on publication of the markdown, but which this "preview" feature here can't show us.